Changing StudentDataHolder to StudentDataFetcher and moving it to br.edu.ufcg.computacao.eureca.backend.core
